Switch back to using the testImplementation scope for Testcontainers
This is actually how the project was initially generated and follows the "Testcontainers at development time" guide.
Although slightly more setup needs to be done, the benefit is that no Testcontainer dependencies are present in production builds.
This reduces the size of the bundle about 25MB and prevents a lot of code that is able to interact with Docker being present on a production system.